refactor: extract archive DB logic into archive_db.py
Moves open_archive_db, check_db_health, contacts/groups persistence,
file tracking, folder sync, and group folder resolution out of
wa_media_archiver.py into a new companion module. Main script reduced
from ~1420 to ~1146 lines. All call sites updated to archive_db.*
and tests updated to load and reference the new module as `arc`.
